[MERGE] 2469-2500

+++ SubtextSolution/Subtext.Framework/ImportExport/ObjectHydrator.cs
@@ -45,8 +45,8 @@ namespace Subtext.ImportExport
 			bmlPost.Content.Text = entry.Body;
 			bmlPost.DateCreated = entry.DateCreated;
 			bmlPost.DateModified = entry.DateModified;
-			bmlPost.PostType = (entry.PostType == PostType.Story) ? BlogPostTypes.Article : BlogPostTypes.Normal;
-			bmlPost.Views = (uint)0; // I think we have this statistic in the db... right?
+            bmlPost.PostType = (entry.PostType == PostType.Story) ? BlogPostTypes.Article : BlogPostTypes.Normal;
+            bmlPost.Views = 0; // I think we have this statistic in the db... right?
 
 			if (entry.HasEntryName)
 			{
